requirements.txt
Flask
requests
google-generativeai
PyJWT